+++ src/app/components/modals/file-chunks.js
@@ -222,7 +222,7 @@ export default Component.extend(PromiseLoadingMixin, {
     if (fileTransfers) {
       fileTransfers.forEach(transfer => {
         if (get(transfer, 'type') === 'invalidation') {
-          providers.add(get(transfer, 'migrationSource'));
+          providers.add(get(transfer, 'invalidatingProvider'));
         }
       });
     }
@@ -339,13 +339,16 @@ export default Component.extend(PromiseLoadingMixin, {
   /**
    * @type {Ember.ComputedProperty<Array<string>>|null}
    */
-  currentMigrationSourceIds: computed('fileTransfers.@each.migrationSource', function () {
-    /** @type {Ember.Array|undefined} */
-    const fileTransfers = this.get('fileTransfers');
-    if (fileTransfers) {
-      return fileTransfers.map(t => get(t, 'migrationSource')).filter(s => s);
+  currentMigrationSourceIds: computed(
+    'fileTransfers.@each.invalidatingProvider',
+    function () {
+      /** @type {Ember.Array|undefined} */
+      const fileTransfers = this.get('fileTransfers');
+      if (fileTransfers) {
+        return fileTransfers.map(t => get(t, 'invalidatingProvider')).filter(s => s);
+      }
     }
-  }),
+  ),
   
   /**
    * Type of element that is presented in modal
@@ -559,9 +562,8 @@ export default Component.extend(PromiseLoadingMixin, {
       const transfer = this.get('store')
         .createRecord('transfer', {
           file,
-          migration: true,
-          migrationSource: source,
-          destination: destination,
+          invalidatingProvider: source,
+          replicatingProvider: destination,
         });
       return transfer.save()
         // TODO: test it and make better fail messages
@@ -590,8 +592,7 @@ export default Component.extend(PromiseLoadingMixin, {
       const transfer = this.get('store')
         .createRecord('transfer', {
           file,
-          migration: false,
-          destination,
+          replicatingProvider: destination,
         });
       return transfer.save()
         .catch(error => {
@@ -619,8 +620,7 @@ export default Component.extend(PromiseLoadingMixin, {
       const transfer = this.get('store')
         .createRecord('transfer', {
           file,
-          migration: true,
-          migrationSource: source,
+          invalidatingProvider: source,
         });
       return transfer.save()
         .catch(error => {